relocate each app and use Router middleware for mini apps integration
